+++ protocols/substreams/ethereum-balancer-v3/src/abi/mod.rs
@@ -1,5 +1,6 @@
 #![allow(clippy::all)]
 
+pub mod quantamm_weighted_pool_factory_contract;
 pub mod reclamm_pool_factory_contract;
 pub mod stable_pool_contract;
 pub mod stable_pool_factory_contract;